made tests modular, removed mvp and symbolic terms
